Feature: Add informations about who is connecting to both sides
- The client will now get the info which dapp he is connecting to.
  Therefore the DApp need to add some more information (IDAppInfos)

- The DApp will get the info what wallet is connecting to it.
  IWalletInfo has the new informations.

- Upon establishing the connection to the DApp, then refresing the
wallet, the server would still have the client connection registered.
But the client did not get any information about that already
established connection. The connect callback now delivers more
information back to the client. On initial connect and on reconnect.

ToDos:
1. Add callbacks telling the client the server connection was
closed. E.g. when the site with the dapp was closed to free up memory
and the connection.

2. Add callback for disconnect, telling the server the client
disconnected and free up all things.

3. Add secure option to allow connection to only secure (wss) server.
   Maybe has to be done in meerkat.