feat: Provide connect status callbacks for client and server
The server has now a new 'disconnect' endpoint, that a client can
trigger before he disconnects.

The server can also trigger a 'shutdown' call to the client. This is
best used when the tab of the browser window, running the DApp is
closed. So the client knows he is no longer connected and can react to
that change.

The client will now get back the address in the connect callback. This
way the client address are on the server and on the client and the
connection can be verified by comparing these connection strings.