feat: Add lifecycle callbacks & subscriber info of the connection
This commit adds information about the two participants of a
peer-to-peer connection. And adds callbacks to inform the other
participant about events like connect, disconnect, and shutdown.

Added types here are IConnectMessage, IDAppInfos, and IWalletInfo.
These contain the minimum viable information needed to identify a
participant for the end user.

IDAppInfos lets the DApp provider that wants to inject the wallet
specify information about its name, URL, and address. This
information gets transmitted to the client once a connection is
created. A wallet provider can show this information to the user
so he knows to which DApp his wallet is connected.

IWalletInfo is the pendant to IDAppInfos. It lets the wallet
specify information like name, version, and icon. This
information is sent to the server when establishing a
connection. This infos already existed in the implementation
but now have been put into an interface.

IConnectMessage is the type that the connect callbacks on the
server and the client will use to have instant information
about the connection status. It includes the info if the
connection is created if there was an error, and if so, it
consists of an optional error message.
It contains the connecting client's address and the IDAppInfos
info for the connected DApp.

The DAppPeerConnect (server) has a new endpoint for letting the
client signal that it is disconnecting from the server.

The DAppPeerConnect connect endpoint will inform the client
connect-callback about:
The user rejected the wallet injection
A different wallet is already connected
The connection has successfully been made

The CardanoPeerConnect (client) has a new endpoint that the server
can call if it shuts down. As a server is only a tab in the browser,
the user might accidentally close it.
The DApp shall call the shutdown function of the server when the
window is closing, so the client can terminate the connection and
inform the user.