feat: Add identicon
The function getIdenticon was added to client and server.

With this function you can get a base64 encoded image of the current
connection.

This image contains 7 charaters from [A-Z0-9] to quickly
identify if the connection that has been establish is the same for
server and client. In addtion the text color has 3 different colors
and in the background different color gradients that make it more easy
for the eye to identify the similarity.

With this we have 3*36^7, roughtly 235 billion combinations with only
the displayed text which should make collisions very unlikely yet make
it easy to identify that the connection is made with the correct endpoint.