feat: Add bidirectional connection
- Wallet and DApp contain a second "discovery" meerkat instance.

Once the wallet has connected, it will tell the server it's discovery
meerkat address. The DApp will store this address.

When the DApp starts again, it will try to connect to this server. If
it connects, it will tell the Wallet to do a normal connect.

On the DApp this feature can be disabled.

-  A temporary InfoLogger has been added, to control the amount of
logging that is happening. This needs to be replaced with a full
logger implementation.