feat: Allow api overwrite
When injecting the api, the caller (the wallet) can overwrite the
default behaviour and replace the api event if it has already been
injected.

This makes it much easier for the wallet to switch between different
wallets. So when changing the used wallet, up until now the wallet had
to disconnect and connect back again, even though the connection was
already established.